issue-711 - Fix hero.css lint
