+++ blocks/ordering-options/ordering-options.js
@@ -1,15 +1,10 @@
-import {
-  detectStore,
-  getCartItemCount,
-  getCartDetails,
-  getOrderingOptions,
-  updateCounters,
-} from '../../scripts/scripts.js';
+import { detectStore, getCartItemCount, setCookie } from '../../scripts/scripts.js';
 import {
   a, div, domEl, h3, i, p, span,
 } from '../../scripts/dom-helpers.js';
 
 const SHOP_BASE_URL = 'https://shop.moleculardevices.com';
+const COOKIE_NAME_CART_ITEM_COUNT = 'cart-item-count';
 
 function increaseAndDecreaseCounter(event) {
   const btnContainer = event.target.closest('span');
@@ -20,6 +15,37 @@ function increaseAndDecreaseCounter(event) {
   } else {
     counterEl.textContent = (counter > 1) ? counter - 1 : 1;
   }
+}
+
+async function updateCounters() {
+  const count = getCartItemCount();
+  const cartCounters = document.querySelectorAll('.cart-count');
+  if (cartCounters) {
+    cartCounters.forEach((cartCounter) => {
+      cartCounter.textContent = count;
+    });
+  }
+}
+
+async function getCartDetails() {
+  return new Promise((resolve) => {
+    const script = domEl('script',
+      {
+        src: `${SHOP_BASE_URL}/cart.json?callback=cartDetails`,
+      },
+    );
+
+    /* eslint-disable dot-notation */
+    window['cartDetails'] = (data) => {
+      document.getElementsByTagName('head')[0].removeChild(script);
+      delete window['cartDetails'];
+      setCookie(COOKIE_NAME_CART_ITEM_COUNT, data.item_count || 0);
+      resolve(data);
+    };
+    /* eslint-enable dot-notation */
+
+    document.getElementsByTagName('head')[0].appendChild(script);
+  });
 }
 
 async function addToCart(event) {
@@ -115,8 +141,33 @@ function renderCartWidget() {
   );
 }
 
-async function renderList(refs, showStore, container) {
-  const options = await getOrderingOptions(refs);
+function fetchOption(option) {
+  return fetch(`${SHOP_BASE_URL}/products/${option}.js`, {
+    mode: 'cors',
+  }).then((response) => {
+    if (response.ok) {
+      return response.json();
+    }
+    return Promise.reject(response);
+  }).catch((err) => {
+    // eslint-disable-next-line no-console
+    console.warn(`Could not fetch ordering details for option ${option}, got status ${err.status}.`, err.statusText);
+  });
+}
+
+async function fetchOptionIntoArray(array, idx, option) {
+  array[idx] = await fetchOption(option.trim());
+}
+
+async function getOrderingOptions(refs) {
+  const options = new Array(refs.length);
+  await Promise.all(refs
+    .map((option, idx) => fetchOptionIntoArray(options, idx, option)),
+  );
+  return options;
+}
+
+async function renderList(options, showStore, container) {
   const items = [];
   options.forEach((option) => {
     items.push(renderItem(option, showStore));
@@ -138,6 +189,178 @@ async function renderList(refs, showStore, container) {
   });
 }
 
+function buildOrderingForm(options) {
+  const orderContainer = document.querySelector('.order-container');
+  if (!orderContainer) return;
+  const optionTitles = options.map((option) => option.title);
+  let variants = [];
+  let selectedOption = null;
+  let selectedVariant = null;
+
+  const orderFormContainer = document.createElement('div');
+  orderFormContainer.classList.add('order-container');
+  orderContainer.appendChild(orderFormContainer);
+
+  function updateVariantsDropdownLabel() {
+    const variantDropDown = document.querySelector('#variantDropDown');
+    if (variantDropDown) {
+      variantDropDown.innerHTML = selectedVariant.title;
+    }
+  }
+
+  function updateDropdownInnerHTML() {
+    const optionsDropdown = document.querySelector('#optionsDropDown');
+    if (optionsDropdown) {
+      optionsDropdown.innerHTML = selectedOption.title;
+    }
+  }
+
+  function handleVariantSelection(variant) {
+    selectedVariant = variant;
+    updateVariantsDropdownLabel();
+    const priceContent = document.querySelector('.price');
+    priceContent.innerHTML = `$ ${(variant.price / 100).toLocaleString('en-US')}.00`;
+  }
+
+  function checkOptionValidity() {
+    if (selectedOption === 'Product Options') {
+      const variantDropDown = document.getElementById('variantDropDown');
+      variantDropDown.classList.toggle('not-allowed');
+      selectedVariant = 'Select Variant';
+      updateVariantsDropdownLabel();
+    } else if (selectedOption !== 'Product Options') {
+      const variantDropDown = document.getElementById('variantDropDown');
+      variantDropDown.classList.add('allowed');
+    }
+  }
+
+  function openDropdownMenu(dropdownId) {
+    const dropdown = document.getElementById(dropdownId);
+    const optionsDropdown = document.querySelector('#optionsDropDown');
+    if (dropdown && optionsDropdown.innerHTML !== 'Product Options') {
+      dropdown.classList.toggle('show');
+      if (dropdownId === 'variantsDropdown') {
+        dropdown.style.left = '550px';
+      }
+    } else if (dropdownId === 'optionsDropdownContent') {
+      dropdown.classList.toggle('show');
+    }
+    checkOptionValidity();
+  }
+
+  function handleOptionSelection(option) {
+    selectedOption = option;
+    updateDropdownInnerHTML();
+    checkOptionValidity();
+    const variantsContent = document.querySelector('#variantsDropdown');
+    variantsContent.replaceChildren();
+    for (let j = 0; j < option.variants.length; j += 1) {
+      const variant = document.createElement('a');
+      variant.innerHTML = option.variants[j].title;
+      variant.classList.add('option');
+      variant.addEventListener('click', () => handleVariantSelection(option.variants[j]));
+      variantsContent.appendChild(variant);
+    }
+  }
+
+  window.onclick = function CloseDropDownMenu(event) {
+    if (!event.target.matches('.drop-down')) {
+      const dropdowns = document.getElementsByClassName('product-options-content');
+      for (let k = 0; k < dropdowns.length; k += 1) {
+        const openDropdown = dropdowns[k];
+        if (openDropdown.classList.contains('show')) {
+          openDropdown.classList.remove('show');
+        }
+      }
+    }
+  };
+
+  // Options dropdown
+  const optionsDropdown = document.createElement('button');
+  optionsDropdown.innerHTML = 'Product Options';
+  optionsDropdown.id = 'optionsDropDown';
+  optionsDropdown.onclick = () => openDropdownMenu('optionsDropdownContent');
+  optionsDropdown.classList.add('drop-down');
+  orderFormContainer.appendChild(optionsDropdown);
+
+  const optionsContent = document.createElement('div');
+  optionsContent.classList.add('product-options-content');
+  optionsContent.id = 'optionsDropdownContent';
+  orderFormContainer.appendChild(optionsContent);
+  for (let l = 0; l < optionTitles.length; l += 1) {
+    const option = document.createElement('a');
+    option.innerHTML = optionTitles[l];
+    option.classList.add('option');
+    option.addEventListener('click', () => handleOptionSelection(options[l]));
+    optionsContent.appendChild(option);
+  }
+  // Variants dropdown
+  const variantDropDown = document.createElement('button');
+  variantDropDown.innerHTML = 'Select Variation';
+  variantDropDown.id = 'variantDropDown';
+  variantDropDown.onclick = () => openDropdownMenu('variantsDropdown');
+  variantDropDown.classList.add('drop-down');
+  orderFormContainer.appendChild(variantDropDown);
+
+  const variantsContent = document.createElement('div');
+  variantsContent.classList.add('product-options-content');
+  variantsContent.id = 'variantsDropdown';
+  orderFormContainer.appendChild(variantsContent);
+
+  const priceLabel = document.createElement('label');
+  priceLabel.classList.add('price-label');
+  priceLabel.innerHTML = 'PRICE';
+  orderFormContainer.appendChild(priceLabel);
+
+  const quantityLabel = document.createElement('label');
+  quantityLabel.classList.add('quantity-label');
+  quantityLabel.innerHTML = 'QUANTITY';
+  orderFormContainer.appendChild(quantityLabel);
+
+  const price = document.createElement('span');
+  price.classList.add('price');
+  price.innerHTML = '$ 0.00';
+  orderFormContainer.appendChild(price);
+
+  const quantityContainer = document.createElement('div');
+  quantityContainer.classList.add('quantity-container');
+  orderFormContainer.appendChild(quantityContainer);
+
+  const decreaseButton = document.createElement('a');
+  decreaseButton.classList.add('quantity-button');
+  decreaseButton.innerHTML = '-';
+  quantityContainer.appendChild(decreaseButton);
+
+  const quantityNumber = document.createElement('span');
+  quantityNumber.classList.add('quantity-number');
+  quantityNumber.innerHTML = '1';
+  quantityContainer.appendChild(quantityNumber);
+
+  const increaseButton = document.createElement('a');
+  increaseButton.classList.add('quantity-button');
+  increaseButton.innerHTML = '+';
+  quantityContainer.appendChild(increaseButton);
+
+  increaseButton.addEventListener('click', () => {
+    let currentQuantity = parseInt(quantityNumber.innerHTML, 10);
+    currentQuantity += 1;
+    quantityNumber.innerHTML = currentQuantity;
+  });
+
+  decreaseButton.addEventListener('click', () => {
+    let currentQuantity = parseInt(quantityNumber.innerHTML, 10);
+    if (currentQuantity > 1) {
+      currentQuantity -= 1;
+      quantityNumber.innerHTML = currentQuantity;
+    }
+  });
+  const addToCartButton = document.createElement('button');
+  addToCartButton.addEventListener('click', () => addToCart(selectedVariant, quantityNumber));
+  addToCartButton.classList.add('add-to-cart');
+  addToCartButton.innerHTML = 'Add to cart';
+  orderFormContainer.appendChild(addToCartButton);
+}
+
 export default async function decorate(block) {
   const refs = [...block.querySelectorAll('.ordering-options > div > div')]
     .map((ref) => (ref.innerHTML).split(', '))
@@ -149,8 +372,10 @@ export default async function decorate(block) {
   block.append(container);
 
   const showStore = detectStore();
-  renderList(refs, showStore, container);
-
+  const orderingOptions = await getOrderingOptions(refs);
+  renderList(orderingOptions, showStore, container);
+  const options = orderingOptions.filter((o) => !!o);
+  buildOrderingForm(options);
   if (showStore) {
     block.classList.add('cart-store');
     await getCartDetails();
